"Solving" the problem with `init`
The issue with init is due to T:ManagedObject
Because initializers return the instance of current class, Swift needs to know what that class (i.e. data type) is, in order to enforce Type Safety.
That means, in order to allow T:ManagedObject.init(), ManagedObject must not be sub-classable, it must be marked as `final`
Which it can't be because we need to create Artist, Album, Track as subclasses of ManagedObject
So, remove the requirement that T is ManagedObject and voila - T.init is now ok

(but then we have new problems)